refactor(signals): extract SignalWidget data pipelines into hooks
Pull the /query/run effect (with its stale-response cancellation guard,
results/errors/rejectStats/latency), the /query/pairs effect, and the
in-browser derive/highlight evaluation out of SignalWidget into
useRunSeries / usePairsData / useDerived (components/signals/useSignalData.ts).
The two fetch effects keep one shared loading flag (useRunSeries owns it,
usePairsData drives it), and effect deps + cancellation semantics are
unchanged. Pure mechanical extraction, no behavior change.
